+++ ImageCropper/src/main/java/com/akshayashokcode/imagecropper/CropShape.kt
@@ -1,0 +1,6 @@
+package com.akshayashokcode.imagecropper
+
+sealed class CropShape {
+    object Rectangle : CropShape()
+    object Circle : CropShape()
+}